move IsSNOCheckFnc function to a common place
